requirements.txt
streamlit==1.32.2
requests==2.31.0

textblob==0.17.1
nltk==3.8.1

langchain==0.1.14
openai==1.14.3

python-dotenv==1.0.1

deep-translator==1.11.4
gtts==2.5.1

scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4